# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Notted API",
    "version": "1.0.0",
    "description": "The versioned public integration surface of Notted. Every path below is served under the `/api/v1` prefix and is covered by the compatibility promise of that version. tRPC is the first-party transport for the Notted web client only: it is an internal implementation detail, is not part of this public contract, and may change without a version bump. Workspace-scoped routes accept an API key bearer token; the session-only routes declare `security: []` and are reachable with a browser session alone."
  },
  "security": [
    {
      "apiKey": []
    }
  ],
  "paths": {
...
    "/api/v1/workspaces/{workspaceId}/api-keys": {
      "get": {
        "summary": "List the workspace's API keys.",
        "tags": [
          "API keys"
        ],
        "description": "Never returns the secret or its hash; only the display prefix.",
        "parameters": [
          {
            "name": "workspaceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": -9007199254740991,
                  "maximum": 9007199254740991
                },
                {
                  "type": "string",
                  "pattern": "^(0|[1-9]\\d*)$"
                }
              ]
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": -9007199254740991,
                  "maximum": 9007199254740991
                },
                {
                  "type": "string",
                  "pattern": "^(0|[1-9]\\d*)$"
                }
              ]
            }
          },
          {
            "name": "includeRevoked",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "true",
                "false"
              ]
            }
          },
          {
            "name": "sortBy",
            "in": "query",
            "required": false,
            "schema": {
              "default": "createdAt",
              "type": "string",
              "enum": [
                "createdAt",
                "lastUsedAt",
                "name"
              ]
            }
          },
          {
            "name": "sortDirection",
            "in": "query",
            "required": false,
            "schema": {
              "default": "desc",
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          }
        ],
        "responses": {
          "2XX": {
            "description": "List the workspace's API keys.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiKeyPage"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Issue an API key.",
        "tags": [
          "API keys"
        ],
        "description": "The only response that carries the raw secret. It is shown exactly once.",
        "parameters": [
          {
            "name": "workspaceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateApiKey"
              }
            }
          }
        },
        "responses": {
          "2XX": {
            "description": "Issue an API key.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiKeyCreateResult"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/workspaces/{workspaceId}/api-keys/{apiKeyId}": {
      "delete": {
        "summary": "Revoke an API key.",
        "tags": [
          "API keys"
        ],
        "parameters": [
          {
            "name": "workspaceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "apiKeyId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "2XX": {
            "description": "Revoke an API key.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiKeyRevokeResult"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "apiKey": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "ntd_pk_*",
        "description": "Workspace API key, sent as `Authorization: Bearer ntd_pk_...`. A key is bound to one workspace and carries `read`, `write`, or `admin` scopes; a request is authorized only when the key's workspace matches the route's `workspaceId` and its scopes cover the action. The secret is returned once, when the key is created."
      }
    },
    "schemas": {
...
      "ApiKeyCreateResult": {
        "type": "object",
        "properties": {
          "apiKey": {
            "$ref": "#/components/schemas/ApiKeySummary"
          },
          "secret": {
            "$ref": "#/components/schemas/ApiKeySecret"
          }
        },
        "required": [
          "apiKey",
          "secret"
        ],
        "additionalProperties": false
      },
      "ApiKeyName": {
        "type": "string",
        "minLength": 1,
        "maxLength": 100
      },
      "ApiKeyPage": {
        "type": "object",
        "properties": {
          "items": {
            "readOnly": true,
            "maxItems": 100,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ApiKeySummary"
            }
          },
          "page": {
            "type": "integer",
            "minimum": 1,
            "maximum": 9007199254740991
          },
          "limit": {
            "type": "integer",
            "minimum": 1,
            "maximum": 100
          },
          "hasMore": {
            "type": "boolean"
          }
        },
        "required": [
          "items",
          "page",
          "limit",
          "hasMore"
        ],
        "additionalProperties": false
      },
      "ApiKeyRevokeResult": {
        "type": "object",
        "properties": {
          "apiKeyId": {
            "$ref": "#/components/schemas/Uuid"
          },
          "revoked": {
            "type": "boolean",
            "const": true
          }
        },
        "required": [
          "apiKeyId",
          "revoked"
        ],
        "additionalProperties": false
      },
      "ApiKeyScope": {
        "type": "string",
        "enum": [
          "read",
          "write",
          "admin"
        ]
      },
      "ApiKeyScopes": {
        "minItems": 1,
        "maxItems": 3,
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/ApiKeyScope"
        }
      },
      "ApiKeySecret": {
        "type": "string",
        "pattern": "^ntd_pk_[A-Za-z0-9_-]{32}$"
      },
      "ApiKeySummary": {
        "type": "object",
        "properties": {
          "id": {
            "$ref": "#/components/schemas/Uuid"
          },
          "workspaceId": {
            "$ref": "#/components/schemas/Uuid"
          },
          "name": {
            "$ref": "#/components/schemas/ApiKeyName"
          },
          "keyPrefix": {
            "type": "string",
            "minLength": 1,
            "maxLength": 8
          },
          "scopes": {
            "$ref": "#/components/schemas/ApiKeyScopes"
          },
          "lastUsedAt": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/IsoTimestamp"
              },
              {
                "type": "null"
              }
            ]
          },
          "expiresAt": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/IsoTimestamp"
              },
              {
                "type": "null"
              }
            ]
          },
          "isRevoked": {
            "type": "boolean"
          },
          "createdById": {
            "$ref": "#/components/schemas/Uuid"
          },
          "createdAt": {
            "$ref": "#/components/schemas/IsoTimestamp"
          }
        },
        "required": [
          "id",
          "workspaceId",
          "name",
          "keyPrefix",
          "scopes",
          "lastUsedAt",
          "expiresAt",
          "isRevoked",
          "createdById",
          "createdAt"
        ],
        "additionalProperties": false
      },
...
      "CreateApiKey": {
        "type": "object",
        "properties": {
          "name": {
            "$ref": "#/components/schemas/ApiKeyName"
          },
          "scopes": {
            "default": [
              "read",
              "write"
            ],
            "$ref": "#/components/schemas/ApiKeyScopes"
          },
          "expiresAt": {
            "$ref": "#/components/schemas/IsoTimestamp"
          }
        },
        "required": [
          "name"
        ],
        "additionalProperties": false
      },
...
      "IsoTimestamp": {
        "type": "string",
        "format": "date-time",
        "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
      },
...
      "Uuid": {
        "type": "string",
        "format": "uuid",
        "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
      },
...
    }
  }
}